transfered exceptionValidation to RepositoryBase
cleanup/refactoring
