+++ ScrumPoker.Infrastructure/DependencyInjection.cs
@@ -4,7 +4,6 @@ using ScrumPoker.Business.Interfaces.Interfaces;
 using ScrumPoker.DataAccess.Data;
 using ScrumPoker.DataAccess.Interfaces;
 using ScrumPoker.DataAccess.Models.EFContext;
-using ScrumPoker.DataAccess.Validation;
 
 namespace ScrumPoker.Infrastructure;
 
@@ -17,7 +16,6 @@ public static class DependencyInjection
         services.AddTransient<IPlayerRepository, PlayerRepository>();
         services.AddTransient<IPlayerService, PlayerService>();
         services.AddTransient<IScrumPokerContext, ScrumPokerContext>();
-        services.AddTransient<IValidation, Validation>();
         services.AddTransient<IVoteRegistrationRepository, VoteRegistrationRepository>();
         services.AddTransient<IRoundRepository, RoundRepository>();
         services.AddTransient<IRoundService, RoundService>();
